layout: Integrate `::first-letter` support into the `InlineFormattingContextBuilder`
This change merges the first letter implementation into the
`InlineFormattingContextBuilder` as its more associated with inline
layout. The downside is that, due to ownership issues in Rust, the
builder must be unwrapped after ensuring it.

This change also makes a few small cleanups to the `first_letter_range`
function in addition to moving it to the inline code.
