+++ components/layout/flow/construct.rs
@@ -11,7 +11,6 @@ use style::properties::longhands::list_style_position::computed_value::T as List
 use style::selector_parser::PseudoElement;
 use style::str::char_is_whitespace;
 use style::values::specified::box_::DisplayOutside as StyloDisplayOutside;
-use unicode_categories::UnicodeCategories;
 
 use super::OutsideMarker;
 use super::inline::construct::InlineFormattingContextBuilder;
@@ -365,89 +364,6 @@ impl<'dom, 'style> BlockContainerBuilder<'dom, 'style> {
     }
 }
 
-/// Computes the range of the first letter.
-///
-/// The range includes any preceding punctuation, and any spaces interleaved
-/// within the preceding punctuation or between the preceding punctuation
-/// and the first letter/number/symbol.
-/// Succeeding punctuation are included in the range, but any space
-/// following the letter/number/symbol ends the range. Intervening
-/// succeeding spaces are not supported yet.
-///
-/// <https://drafts.csswg.org/css-pseudo/#first-letter-pattern>
-fn first_letter_range(text: &str) -> std::ops::Range<usize> {
-    use unicode_categories::UnicodeCategories;
-
-    enum State {
-        Start,
-        PrecedingPunc,
-        /// Unicode general category L: letter, N: number and S: symbol
-        Lns,
-        SucceedingPunc,
-    }
-
-    let mut start = 0;
-    let mut end = None;
-    let mut state = State::Start;
-
-    for (i, c) in text.char_indices() {
-        match &mut state {
-            State::Start => {
-                if c.is_punctuation() {
-                    start = i;
-                    state = State::PrecedingPunc;
-                } else if c.is_letter() || c.is_number() || c.is_symbol() {
-                    start = i;
-                    state = State::Lns;
-                } else if c.is_separator_space() {
-                    continue;
-                } else {
-                    // Found invalid character
-                    return 0..0;
-                }
-            },
-            State::PrecedingPunc => {
-                if c.is_letter() || c.is_number() || c.is_symbol() {
-                    state = State::Lns;
-                } else if c.is_punctuation() || (c.is_separator_space() && c != '\u{3000}') {
-                    continue;
-                } else {
-                    // Found invalid character
-                    return 0..0;
-                }
-            },
-            State::Lns => {
-                // TODO: Implement support for intervening spaces
-                // <https://drafts.csswg.org/css-pseudo/#first-letter-pattern>
-                if c.is_punctuation() && !c.is_punctuation_open() && !c.is_punctuation_dash() {
-                    state = State::SucceedingPunc;
-                } else {
-                    end = Some(i);
-                    break;
-                }
-            },
-            State::SucceedingPunc => {
-                // TODO: Implement support for intervening spaces
-                // <https://drafts.csswg.org/css-pseudo/#first-letter-pattern>
-                if c.is_punctuation() && !c.is_punctuation_open() && !c.is_punctuation_dash() {
-                    continue;
-                } else {
-                    end = Some(i);
-                    break;
-                }
-            },
-        }
-    }
-
-    match state {
-        State::Start | State::PrecedingPunc => 0..0,
-        State::Lns | State::SucceedingPunc => {
-            let end = end.unwrap_or(text.len());
-            start..end
-        },
-    }
-}
-
 impl<'dom> TraversalHandler<'dom> for BlockContainerBuilder<'dom, '_> {
     fn handle_element(
         &mut self,
@@ -508,42 +424,11 @@ impl<'dom> TraversalHandler<'dom> for BlockContainerBuilder<'dom, '_> {
             self.finish_anonymous_table_if_needed();
         }
 
-        let container_info = self.info;
-        let context = self.context;
-        let builder = self.ensure_inline_formatting_context_builder();
-
-        // ::first-letter is an eager pseudo element and should not be nested
-        if let Some(pseudo_info) = (container_info.pseudo_element_chain().is_empty())
-            .then(|| container_info.with_pseudo_element(context, PseudoElement::FirstLetter))
-            .flatten()
-            .filter(|_| {
-                builder
-                    .text_segments
-                    .iter()
-                    .flat_map(|seg| seg.chars())
-                    .all(|c| c.is_separator_space())
-            })
-        {
-            let first_letter_range = first_letter_range(&text[..]);
-
-            // The first letter range may be some value larger than zero when
-            // there are preceding spaces.
-            if first_letter_range.start != 0 {
-                builder.push_text(Cow::Borrowed(&text[0..first_letter_range.start]), info);
-            }
-
-            builder.start_inline_box(
-                || ArcRefCell::new(InlineBox::new(&pseudo_info, context)),
-                None,
-            );
-            let first_letter_text = Cow::Borrowed(&text[first_letter_range.clone()]);
-            builder.push_text(first_letter_text, &pseudo_info);
-            builder.end_inline_box();
-
-            builder.push_text(Cow::Borrowed(&text[first_letter_range.end..]), info);
-        } else {
-            builder.push_text(text, info);
-        }
+        self.ensure_inline_formatting_context_builder();
+        self.inline_formatting_context_builder
+            .as_mut()
+            .expect("Should be guaranteed by line above")
+            .push_text_with_possible_first_letter(text, info, self.info, self.context);
     }
 
     fn enter_display_contents(&mut self, styles: SharedInlineStyles) {
@@ -635,13 +520,11 @@ impl<'dom> BlockContainerBuilder<'dom, '_> {
         // Otherwise, this is just a normal inline box. Whatever happened before, all we need to do
         // before recurring is to remember this ongoing inline level box.
         let inline_builder = self.ensure_inline_formatting_context_builder();
-        inline_builder.start_inline_box(
+        let inline_item = inline_builder.start_inline_box(
             || ArcRefCell::new(InlineBox::new(info, context)),
             old_layout_box,
         );
-        box_slot.set(LayoutBox::InlineLevel(
-            inline_builder.inline_items.last().unwrap().clone(),
-        ));
+        box_slot.set(LayoutBox::InlineLevel(inline_item));
 
         if is_list_item {
             if let Some((marker_info, marker_contents)) =
@@ -918,78 +801,5 @@ impl IntermediateBlockContainer {
             } => BlockContainer::construct(context, info, contents, propagated_data, is_list_item),
             IntermediateBlockContainer::InlineFormattingContext(block_container) => block_container,
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn assert_first_letter_eq(text: &str, expected: &str) {
-        let range = first_letter_range(text);
-        assert_eq!(&text[range], expected);
-    }
-
-    #[test]
-    fn test_first_letter_range() {
-        // All spaces
-        assert_first_letter_eq("", "");
-        assert_first_letter_eq("  ", "");
-
-        // Spaces and punctuation only
-        assert_first_letter_eq("(", "");
-        assert_first_letter_eq(" (", "");
-        assert_first_letter_eq("( ", "");
-        assert_first_letter_eq("()", "");
-
-        // Invalid chars
-        assert_first_letter_eq("\u{0903}", "");
-
-        // First letter only
-        assert_first_letter_eq("A", "A");
-        assert_first_letter_eq(" A", "A");
-        assert_first_letter_eq("A ", "A");
-        assert_first_letter_eq(" A ", "A");
-
-        // Word
-        assert_first_letter_eq("App", "A");
-        assert_first_letter_eq(" App", "A");
-        assert_first_letter_eq("App ", "A");
-
-        // Preceding punctuation(s), intervening spaces and first letter
-        assert_first_letter_eq(r#""A"#, r#""A"#);
-        assert_first_letter_eq(r#" "A"#, r#""A"#);
-        assert_first_letter_eq(r#""A "#, r#""A"#);
-        assert_first_letter_eq(r#"" A"#, r#"" A"#);
-        assert_first_letter_eq(r#" "A "#, r#""A"#);
-        assert_first_letter_eq(r#"("A"#, r#"("A"#);
-        assert_first_letter_eq(r#" ("A"#, r#"("A"#);
-        assert_first_letter_eq(r#"( "A"#, r#"( "A"#);
-        assert_first_letter_eq(r#"[ ( "A"#, r#"[ ( "A"#);
-
-        // First letter and succeeding punctuation(s)
-        // TODO: modify test cases when intervening spaces in succeeding puntuations is supported
-        assert_first_letter_eq(r#"A""#, r#"A""#);
-        assert_first_letter_eq(r#"A" "#, r#"A""#);
-        assert_first_letter_eq(r#"A)]"#, r#"A)]"#);
-        assert_first_letter_eq(r#"A" )]"#, r#"A""#);
-        assert_first_letter_eq(r#"A)] >"#, r#"A)]"#);
-
-        // All
-        assert_first_letter_eq(r#" ("A" )]"#, r#"("A""#);
-        assert_first_letter_eq(r#" ("A")] >"#, r#"("A")]"#);
-
-        // Non ASCII chars
-        assert_first_letter_eq("一", "一");
-        assert_first_letter_eq(" 一 ", "一");
-        assert_first_letter_eq("一二三", "一");
-        assert_first_letter_eq(" 一二三 ", "一");
-        assert_first_letter_eq("（一二三）", "（一");
-        assert_first_letter_eq(" （一二三） ", "（一");
-        assert_first_letter_eq("（（一", "（（一");
-        assert_first_letter_eq(" （ （一", "（ （一");
-        assert_first_letter_eq("一）", "一）");
-        assert_first_letter_eq("一））", "一））");
-        assert_first_letter_eq("一） ）", "一）");
     }
 }
