+++ tests/wpt/meta/css/CSS2/selectors/first-letter-dynamic-002.xht.ini
@@ -1,0 +1,2 @@
+[first-letter-dynamic-002.xht]
+  expected: FAIL
