--- tests/wpt/meta/css/CSS2/selectors/first-letter-nested-001.xht.ini
@@ -1,2 +1,0 @@
-[first-letter-nested-001.xht]
-  expected: FAIL
