--- tests/wpt/meta/css/CSS2/selectors/first-letter-nested-002.xht.ini
@@ -1,2 +1,0 @@
-[first-letter-nested-002.xht]
-  expected: FAIL
