+++ tests/wpt/meta/css/CSS2/selectors/first-letter-selector-005.xht.ini
@@ -1,0 +1,2 @@
+[first-letter-selector-005.xht]
+  expected: FAIL
