--- tests/wpt/meta/css/css-grid/grid-model/grid-first-letter-003.html.ini
@@ -1,2 +1,0 @@
-[grid-first-letter-003.html]
-  expected: FAIL
