--- tests/wpt/meta/css/css-grid/grid-model/grid-inline-first-letter-003.html.ini
@@ -1,2 +1,0 @@
-[grid-inline-first-letter-003.html]
-  expected: FAIL
