--- tests/wpt/meta/paint-timing/fcp-only/fcp-typographic-pseudo.html.ini
@@ -1,4 +1,0 @@
-[fcp-typographic-pseudo.html]
-  expected: TIMEOUT
-  [First contentful paint fires only when some of the text is visible, considering ::first-letter.]
-    expected: TIMEOUT
